he cambiado el nombre a initDB y copia ddel main, si se rompe es por eso
